ImpureCode rule split
