+++ src/main/kotlin/pl/setblack/detekt/kurepotlin/ImpureProvider.kt
@@ -4,9 +4,12 @@ import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
 import pl.setblack.detekt.kurepotlin.rules.ClassDefinition
-import pl.setblack.detekt.kurepotlin.rules.ImpureCode
+import pl.setblack.detekt.kurepotlin.rules.LoopUsage
 import pl.setblack.detekt.kurepotlin.rules.MutableCollectionsCode
+import pl.setblack.detekt.kurepotlin.rules.ReturnStatement
+import pl.setblack.detekt.kurepotlin.rules.ReturnUnit
 import pl.setblack.detekt.kurepotlin.rules.ThrowExpression
+import pl.setblack.detekt.kurepotlin.rules.VariableUsage
 
 
 class ImpureProvider : RuleSetProvider {
@@ -16,11 +19,13 @@ class ImpureProvider : RuleSetProvider {
     override fun instance(config: Config): RuleSet = RuleSet(
         ruleSetId,
         listOf(
-            ImpureCode(),
+            LoopUsage(config),
+            ReturnStatement(config),
+            VariableUsage(config),
+            ReturnUnit(config),
             ClassDefinition(config),
             ThrowExpression(config),
             MutableCollectionsCode(config)
-
         )
     )
 }
