+++ src/main/kotlin/pl/setblack/detekt/kurepotlin/rules/ReturnUnit.kt
@@ -1,0 +1,42 @@
+package pl.setblack.detekt.kurepotlin.rules
+
+import io.gitlab.arturbosch.detekt.api.CodeSmell
+import io.gitlab.arturbosch.detekt.api.Config
+import io.gitlab.arturbosch.detekt.api.Debt
+import io.gitlab.arturbosch.detekt.api.Entity
+import io.gitlab.arturbosch.detekt.api.Issue
+import io.gitlab.arturbosch.detekt.api.Rule
+import io.gitlab.arturbosch.detekt.api.Severity
+import org.jetbrains.kotlin.psi.KtLoopExpression
+import org.jetbrains.kotlin.psi.KtNamedFunction
+import org.jetbrains.kotlin.psi.KtProperty
+import org.jetbrains.kotlin.psi.KtReturnExpression
+import org.jetbrains.kotlin.resolve.BindingContext
+import org.jetbrains.kotlin.types.KotlinType
+import org.jetbrains.kotlin.types.typeUtil.isUnit
+
+class ReturnUnit(config: Config = Config.empty) : Rule(config) {
+    override val issue = Issue(
+        javaClass.simpleName,
+        Severity.CodeSmell,
+        "Pure function should return something",
+        Debt.TWENTY_MINS
+    )
+
+    override fun visitNamedFunction(function: KtNamedFunction) {
+        bindingContext.takeIf { it != BindingContext.EMPTY }
+            ?.get(BindingContext.FUNCTION, function)
+            ?.returnType
+            ?.takeIf(KotlinType::isUnit)
+            ?.let {
+                val file = function.containingKtFile
+                report(
+                    CodeSmell(
+                        issue, Entity.from(function),
+                        message = "Function ${function.name} in the file ${file.name} returns Unit."
+                    )
+                )
+            }
+        super.visitNamedFunction(function)
+    }
+}
